reorganized components into separate folders and created individual style files
