+++ src/components/Canva/canva.css
@@ -1,0 +1,9 @@
+.canva {
+    width: 90vw;
+    height: 80vh;
+    background-color: aliceblue;
+    touch-action: none;
+    max-width: 1200px;
+    max-height: 675px;
+    margin-top: 1rem;
+}
