+++ src/components/Canva/components/Box/Box.js
@@ -1,6 +1,7 @@
 import React from "react";
+import './box.css'
 import { observer } from "mobx-react";
-import BoxDraggable from "./BoxDraggable";
+import BoxDraggable from "./components/BoxDraggable/BoxDraggable";
 
 function Box(props) {
   return (
