+++ src/components/Canva/components/Box/box.css
@@ -1,0 +1,6 @@
+.box {
+    position: absolute;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
