+++ src/components/Toolbar/toolbar.css
@@ -1,0 +1,22 @@
+.toolbar {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    padding: 10px;
+    column-gap: 10px;
+    row-gap: 10px;
+    width: 100%;
+    height: auto;
+
+    span {
+        width: max-content;
+        align-content: center;
+        padding: 5px;
+    }
+
+}
+
+.toolbar span {
+    color: white;
+}
