+++ src/main.css
@@ -14,51 +14,16 @@ html,
 
 .app {
     width: 100%;
-    height: 100%;
+    height: auto;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-}
-
-.canva {
-    width: 90vw;
-    height: 80vh;
-    background-color: aliceblue;
-    touch-action: none;
-    max-width: 1200px;
-    max-height: 675px;
-    margin-top: 3rem;
+    padding: 1rem;
 
 }
 
-.box {
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
 
-.toolbar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    padding: 10px;
-    column-gap: 10px;
-    width: 100%;
-    height: auto;
 
-    span {
-        width: max-content;
-        align-content: center;
-        padding: 5px;
-    }
 
-}
 
-.toolbar span {
-    color: white;
-}
