+++ src/tests/App.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "../components/App";
+import App from "../App";
 
 test("Renders correctly the app", () => {
   const div = document.createElement("div");
